파일 확장자 변경 (.js > .ts)
