--- src/config/db.js
@@ -1,5 +1,0 @@
-var Prisma = require('@prisma/client');
-
-var db = new Prisma.PrismaClient();
-
-module.exports = { db };
