+++ src/config/db.ts
@@ -1,0 +1,5 @@
+import Prisma from '@prisma/client';
+
+var db = new Prisma.PrismaClient();
+
+export default db;
