+++ src/types/auth.ts
@@ -1,0 +1,10 @@
+export interface RegisterUser {
+    email: string;
+    nickname: string;
+    password: string;
+}
+
+export interface LoginUser {
+    email: string;
+    password: string;
+}
